Share rotation wait code
